review: Move displayed row IDs out of terminal rendering
Review records now live in data.file_review. The helper that maps stored
selection IDs to the short numeric IDs shown beside changed rows still
lives in terminal output.

Selection commands and rendering both need that mapping, and the
calculation does not format terminal text. Keeping it in output makes
data-oriented selection code depend on presentation.

This commit moves displayed-row ID mapping beside the review records and
updates its selection and rendering callers.

The next commit will update architecture tests before calculations for
selectable review rows move below terminal output.
